VolumeMesher is in 1 .h/.cc file
