--- framework/mesh/VolumeMesher/chi_volumemesher_02_execute.cc
@@ -1,12 +1,0 @@
-#include "chi_volumemesher.h"
-#include <iostream>
-
-#include "chi_runtime.h"
-#include "chi_log.h"
-
-void
-chi_mesh::VolumeMesher::Execute()
-{
-  Chi::log.Log() << "Empty volume mesher, nothing executed.";
-  Chi::log.Log() << std::endl;
-}
